Start on main page - list and add
I started to work on the main page. I started to write the HTML/CSS and style the site a little.
I then coded the list. I fixed it so that all cities are displayed when the site is loaded, and that you can remove cities from the list and see them dissappear without having to refresh the page.
I then made the "Add city" part.
Using the two inputs and a button, it sends a POST requests and adds the City to the list without the need for a refresh. If the city already exists in the database, a status message will tell the user about this.