Search cities + bug fix with "delete" buttons
I coded the "search cities" part of the site, where the user can search for cities. When the button is pressed, a request is sent to the server with the value of the inputs. If the "text" input is empty, an error will be displayed. If no cities are found, the user will be informed about this. If cities are found, they are displayed under the header "Search result:"

I also noticed a bug with my delete-buttons, the cities were not getting deleted and i got the response status "400".
I fixed this by editing the "eventListeners" function, where the buttons could not access the "cityID" key.